requirements.txt
flask 
openai 
python-dotenv
pytube